requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
dvc==3.41.0
dvc-s3==3.1.0
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
mlflow==2.9.2
matplotlib==3.8.2
seaborn==0.13.0
great-expectations==0.18.8

